update: moved testconfigbuilder to tests/config.rs & added docs , drop all collections not just jobs
