+++ crates/orchestrator/src/database/mongodb/mod.rs
@@ -40,6 +40,9 @@ impl MongoDb {
         MongoDb { client }
     }
 
+    /// Mongodb client uses Arc internally, reducing the cost of clone.
+    /// Directly using clone is not recommended for libraries not using Arc internally.
+    /// Dev might want to pass an Arc manually.
     pub fn client(&self) -> Client {
         self.client.clone()
     }
