+++ crates/orchestrator/src/tests/database/mod.rs
@@ -1,6 +1,7 @@
-use crate::config::{config, TestConfigBuilder};
+use crate::config::config;
 use crate::jobs::types::{ExternalId, JobItem, JobStatus, JobType};
 use crate::tests::common::drop_database;
+use crate::tests::config::TestConfigBuilder;
 use rstest::*;
 use uuid::Uuid;
