+++ crates/orchestrator/src/tests/mod.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod database;
 
 pub mod jobs;
